+++ CS16-effie/CS16-effie/MitOutput.swift
@@ -1,0 +1,36 @@
+//
+//  MitOutput.swift
+//  CS16-effie
+//
+//  Created by Effie on 2023/02/27.
+//
+
+import Foundation
+
+protocol MitOutput: CustomStringConvertible {
+  associatedtype InfoType
+  
+  var fileName: String { get }
+  var info: InfoType { get }
+}
+
+struct ListOutput: MitOutput {
+  typealias InfoType = Int
+  
+  var fileName: String
+  var info: Int
+  
+  var description: String {
+    "\(fileName) \(info)KB"
+  }
+}
+
+struct HashOutput: MitOutput {
+  typealias InfoType = String
+  
+  var fileName: String
+  var info: String
+  var description: String {
+    "\(fileName) = \(info)"
+  }
+}
